+++ src/components/Tooltip/Tooltip.types.ts
@@ -20,4 +20,5 @@ export type Props = Pick<
 	// eslint-disable-next-line @typescript-eslint/no-empty-object-type
 	children: (attributes: Parameters<FlyoutTriggerProps["children"]>[0] | {}) => React.ReactNode;
 	text?: React.ReactNode;
+	color?: "inverted" | "dark";
 };